scene: Implement initial animation playback (#240)
* asset: Output skeleton anims

* asset: Output animation data

* scene: Load animation data into skeleton template

* asset: Add default gltf animation channels

* scene: Initial prototype animation playback

* asset: Minor optimization to mesh component size

* scene: Minor naming cleanup in skeleton

* asset: Refactor skeleton InverseBindMatrix storage

* asset: Output anim data as aligned vecs

* scene: Implement scale animation

* asset: Refactor gltf_anim_data_push_access_vec

* asset: Refactor invBindMat computation

* asset: Convert animations from R to L coords

* asset: Parse gltf animation duration

* asset: Minor cleanup in gltf loader

* asset: Minor refactor to gltf data storage

* asset: Removed unused import

* asset: Removed unused parts of gltf loader

Both version and extension information is not used atm.
If we need it we can reimplement this.

* asset: Rename gltf loader json helpers

* asset: Minor refactoring in gltf loader

* asset: Refactor gltf json vector parsing

* asset: Cleanup in gltf loader

* scene: Add seperate animation component

* scene: Initial animation layers

* debug: Initial animation debug panel

* debug: Draw layer time

* debug: Add progress to animation debugger

* scene: Add speed control to anim playback

* ui: Fix missing spacing in table columns

* asset: Rename bind pose mats

* asset: Add seperate default skeleton pose

* scene: Refactor joint sampling

* scene: Minor shortening of skeleton names

* scene: Tweaked anim layer sampling

* scene: Skeleton refactoring

* scene: Refactor skeleton sampling

* asset: Fix uninitialized anim channel

* geo: Add quaternion dot product

* geo: Add quaternion flip util

* scene: Initial anim blending

* scene: Tie layer index into anim index

* scene: Fix anim blending with default pose

+++ libs/asset/src/loader_mesh_pme.c
@@ -146,7 +146,7 @@ static f32 pme_def_axis_scale(const PmeDef* def) {
   diag_crash();
 }
 
-static usize pme_max_verts(PmeDef* def) {
+static u32 pme_max_verts(PmeDef* def) {
   /**
    * Get a conservative maximum amount of needed vertices.
    */
